[PROT-386] Start Implementing scheduler test for Block  fixing constraints TODO:// Check Constraints TODO:// Refactor LibMediaItem to fill LibFileItem TODO:// Implemments test for Building SchEventConfiguration SchClockConfiguration and grid TODO:// Log upload pipe TODO:// Finish merging TODO:// Fix unit test TODO:// Write unit tests

+++ scheduler/src/main/java/io/protone/scheduler/service/SchAttachmentService.java
@@ -1,5 +1,6 @@
 package io.protone.scheduler.service;
 
+import com.google.common.collect.Sets;
 import io.protone.scheduler.domain.SchAttachment;
 import io.protone.scheduler.repository.SchAttachmentRepository;
 import org.slf4j.Logger;
@@ -7,6 +8,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.inject.Inject;
 import java.util.Set;
 
 import static java.util.stream.Collectors.toSet;
@@ -16,16 +18,23 @@ import static java.util.stream.Collectors.toSet;
 public class SchAttachmentService {
 
     private final Logger log = LoggerFactory.getLogger(SchAttachmentService.class);
+
+    @Inject
     private SchAttachmentRepository schAttachmentRepository;
 
     @Transactional
     public Set<SchAttachment> saveAttachmenst(Set<SchAttachment> schAttachmentSet) {
-        return schAttachmentSet.stream().map(schAttachment -> schAttachmentRepository.saveAndFlush(schAttachment)).collect(toSet());
+        if (schAttachmentSet != null && !schAttachmentSet.isEmpty()) {
+            return schAttachmentSet.stream().map(schAttachment -> schAttachmentRepository.saveAndFlush(schAttachment)).collect(toSet());
+        }
+        return Sets.newHashSet();
     }
 
     @Transactional
     public void deleteAttachments(Set<SchAttachment> schAttachmentSet) {
-        schAttachmentSet.stream().forEach(schAttachment -> schAttachmentRepository.delete(schAttachment));
+        if (schAttachmentSet != null && !schAttachmentSet.isEmpty()) {
+            schAttachmentSet.stream().forEach(schAttachment -> schAttachmentRepository.delete(schAttachment));
+        }
     }
 
 }
